weather_mcp/requirements.txt
fastmcp>=2.2.6
PyYAML>=6.0
requests>=2.0
python-dotenv>=0.19.0
